+++ packages/workflow/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rilaykit/workflow
 
+## 0.2.0-beta.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @rilaykit/core@0.2.0-beta.1
+  - @rilaykit/forms@0.2.0-beta.1
+
 ## 0.2.0-beta.0
 
 ### Minor Changes